refactor: Move load tests
